refactor(generator): extract generator to a separate gem
