+++ spec/support/shared_examples/config.yml_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'html2rss/configs/helper'
+require_relative '../helper'
 require 'tzinfo'
 
 RSpec.shared_examples 'config.yml' do |file_name, params|
@@ -20,7 +20,7 @@ RSpec.shared_examples 'config.yml' do |file_name, params|
 
     it "resides in a folder named after channel.url's host" do
       dirname = File.dirname(file.path).split(File::Separator).last
-      host_name = Html2rss::Configs::Helper.url_to_directory_name yaml['channel']['url']
+      host_name = Helper.url_to_directory_name yaml['channel']['url']
 
       expect(dirname).to eq(host_name)
     end
@@ -64,7 +64,7 @@ RSpec.shared_examples 'config.yml' do |file_name, params|
 
       context 'with template post_processor' do
         it 'references available selectors only', aggregate_failures: true do
-          Html2rss::Configs::Helper.referenced_selectors_in_template(yaml['selectors']).each do |referenced_selector|
+          Helper.referenced_selectors_in_template(yaml['selectors']).each do |referenced_selector|
             next if referenced_selector == 'self'
 
             expect(yaml['selectors'][referenced_selector])
